+++ apps/api/src/services/impersonation/impersonation.module.ts
@@ -1,3 +1,4 @@
+import { AccessModule } from '@ghostfolio/api/app/access/access.module';
 import { SubscriptionModule } from '@ghostfolio/api/app/subscription/subscription.module';
 import { ConfigurationModule } from '@ghostfolio/api/services/configuration/configuration.module';
 import { ImpersonationService } from '@ghostfolio/api/services/impersonation/impersonation.service';
@@ -6,7 +7,12 @@ import { PrismaModule } from '@ghostfolio/api/services/prisma/prisma.module';
 import { Module } from '@nestjs/common';
 
 @Module({
-  imports: [ConfigurationModule, PrismaModule, SubscriptionModule],
+  imports: [
+    AccessModule,
+    ConfigurationModule,
+    PrismaModule,
+    SubscriptionModule
+  ],
   providers: [ImpersonationService],
   exports: [ImpersonationService]
 })