+++ libs/common/src/lib/dtos/create-access.dto.ts
@@ -1,20 +1,27 @@
 import { Filter } from '@ghostfolio/common/interfaces';
 import { Scope, scopes } from '@ghostfolio/common/scopes';
+import { IsInTheFutureConstraint } from '@ghostfolio/common/validator-constraints/is-in-the-future';
 
 import { AccessType } from '@prisma/client';
 import {
   IsArray,
+  IsDateString,
   IsEnum,
   IsIn,
   IsOptional,
   IsString,
-  IsUUID
+  IsUUID,
+  Validate
 } from 'class-validator';
 
 export class CreateAccessDto {
   @IsOptional()
   @IsString()
   alias?: string;
+
+  @IsDateString()
+  @Validate(IsInTheFutureConstraint)
+  expiresAt: string;
 
   @IsArray()
   @IsOptional()