+++ libs/common/src/lib/dtos/update-access.dto.ts
@@ -1,12 +1,25 @@
 import { Filter } from '@ghostfolio/common/interfaces';
 import { Scope, scopes } from '@ghostfolio/common/scopes';
+import { IsInTheFutureConstraint } from '@ghostfolio/common/validator-constraints/is-in-the-future';
 
-import { IsArray, IsIn, IsOptional, IsString, IsUUID } from 'class-validator';
+import {
+  IsArray,
+  IsDateString,
+  IsIn,
+  IsOptional,
+  IsString,
+  IsUUID,
+  Validate
+} from 'class-validator';
 
 export class UpdateAccessDto {
   @IsOptional()
   @IsString()
   alias?: string;
+
+  @IsDateString()
+  @Validate(IsInTheFutureConstraint)
+  expiresAt: string;
 
   @IsArray()
   @IsOptional()