+++ libs/common/src/lib/interfaces/access.interface.ts
@@ -4,8 +4,10 @@ import { AccessSettings } from './access-settings.interface';
 
 export interface Access {
   alias: string | null;
+  expiresAt: Date;
   grantee?: string;
   id: string;
+  lastUsedAt?: Date | null;
   scopes: string[];
   settings?: AccessSettings;
   type: AccessType;